# Exp: Add RTree index for polygon