--- a/README.md
+++ b/README.md
@@ -7,8 +7,6 @@ for [ringsaturn/tzf-rs](https://github.com/ringsaturn/tzf-rs).
 cargo add geometry-rs
 ```
 
-For usage example, please see [`./examples/demo.rs`](./examples/demo.rs).
-
 ## TODO
 
 - [ ] Use compressed data of RTree index, since too many memory costs.