--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -31,6 +31,7 @@ mod benches_polygon {
                 },
             ],
             vec![],
+            false,
         );
         return poly;
     }