--- a/benches/bench_az.rs
+++ b/benches/bench_az.rs
@@ -1594,6 +1594,7 @@ mod benches_az_polygon {
                 
             ],
             vec![],
+            true,
         );
         return poly;
     }