--- a/examples/demo.rs
+++ b/examples/demo.rs
@@ -27,6 +27,7 @@ fn main() {
             },
         ],
         vec![],
+        false,
     );
 
     let p_out = geometry_rs::Point {