- Please note that this PR is still WIP.
- Memory usage increase a lot for tzf-ts use case(from 40MB to 140MB), which should be fixed.
- For timezone query in [tzf-rs's `Finder`](https://docs.rs/tzf-rs/0.4.1/tzf_rs/struct.Finder.html) use case, a benchmark runs on world cities on my MacBook Pro:
   - Previous version could run 70 Kops/s
   - This version could run 170 Kops/s
- tzfpy perf diff:
   ```diff
   Name (time in us)               Min      Max    Mean  StdDev  Median     IQR  Outliers  OPS (Kops/s)  Rounds  Iterations
   ------------------------------------------------------------------------------------------------------------------------
   -test_tzfpy_random_cities     2.5024  26.9442  5.7295  2.7556  5.1883  3.3751  6551;618      174.5363   20000          10
   +test_tzfpy_random_cities     2.4133  14.5704  3.8425  0.9685  3.6863  1.2162  6779;364      260.2482   20000          10
   ------------------------------------------------------------------------------------------------------------------------
   ```
- [US AZ benchmark](https://github.com/ringsaturn/geometry-rs/blob/main/benches/bench_az.rs)
  ```diff
  -test benches_az_polygon::poly_contain_point     ... bench:       2,704 ns/iter (+/- 18)
  -test benches_az_polygon::poly_not_contain_point ... bench:       2,715 ns/iter (+/- 5)
  +test benches_az_polygon::poly_contain_point     ... bench:         189 ns/iter (+/- 2)
  +test benches_az_polygon::poly_not_contain_point ... bench:          90 ns/iter (+/- 2)
  ```
- Refer #3

---

TODO:

- [ ] ~~Reduce memory, RTree cost lot.~~ Move to #6 
- [x] Add option to set whether index is need: for most case RTree index is not need.
- [x] Append a real world benchmark for point-in-polygon check: another PR will do this. https://github.com/ringsaturn/geometry-rs/pull/5