@ringsaturn 2023-06-18
For now, performance seems close to original Go repo(https://github.com/tidwall/geojson/issues/23).

Reduce memory usage should be done so this PR could be merged.